+++ app/test/shared/theme/widgets/dovah_button.widget_test.dart
@@ -153,8 +153,8 @@ void main() {
         expect(data.flagsCollection.isButton, isTrue);
         expect(data.flagsCollection.isEnabled, Tristate.isTrue);
         expect(data.hasAction(SemanticsAction.tap), isTrue);
-        tester.binding.pipelineOwner.semanticsOwner!.performAction(
-          node.id,
+        tester.semantics.performAction(
+          find.semantics.byLabel('Confirm'),
           SemanticsAction.tap,
         );
         expect(tapCount, 1);