+++ app/test/shared/theme/widgets/dovah_connection_card.widget_test.dart
@@ -276,8 +276,8 @@ void main() {
         expect(data.flagsCollection.isButton, isTrue);
         expect(data.flagsCollection.isEnabled, Tristate.isTrue);
         expect(data.hasAction(SemanticsAction.tap), isTrue);
-        tester.binding.pipelineOwner.semanticsOwner!.performAction(
-          node.id,
+        tester.semantics.performAction(
+          find.semantics.byLabel(label),
           SemanticsAction.tap,
         );
         expect(tapCount, 1);